fixing MuonStubMakerBase.cc MuonStubMakerBase.cc
in XDigiToStubsConverter::makeStubs, removed
  if (!dtPhDigis) return;
  if (!cscDigis) return;
  if (!rpcDigis) return;
This is not needed as the DigiToStubsConverters are added dynamically to
the vector digiToStubsConverters if given digis are configured to be
used (parameter dropXXPrimitives is false in the config). But it was
confusing: when the digi collection was not in the file, the job was
still running without reporting the missing collection.