possibility of configuring the OMTF emulator directly from XMLs
OMTFReconstruction.h and OMTFReconstruction.cc
The possibility of configuring the OMTF emulator directly from the
config XML is added. When both patternsXMLFile and configXMLFile are
given from the python config of the simOmtfDigis or simOmtfPhase2Digis,
the L1TMuonOverlapParams are read directly from these files, and
the L1TMuonOverlapPhase1ParamsESProducer is not needed. Otherwise, the
L1TMuonOverlapParams are obtained from the eventSetup (so from the DB,
or from the L1TMuonOverlapPhase1ParamsESProducer - if it is defined).

L1Trigger/L1TMuonOverlapPhase2/python/fakeOmtfParamsPhase2_cff.py is
removed, as when it is used, the L1TMuonOverlapPhase1 gets this config
as well.

L1Trigger/L1TMuonOverlapPhase2/python/simOmtfPhase2Digis_extrapol_cfi.py
is moved to
simOmtfPhase2Digis_cfi.py, as this is the default configuration
(besides, there should be only one cfi for a given module).

L1Trigger/L1TMuonOverlapPhase1/src/Omtf/OMTFProcessor.cc
line
trackAddr[3] = myCand->getGpResultUnconstr().getPdfSumUnconstr();
is removed, as uGMT expects 3 entries in the TrackAddress (when there
were 4, uGMT was generating a warning).

L1Trigger/L1TMuonOverlapPhase2/test - runMuonOverlap*.py are corrected